modularizado y arreglo en los comentarios
