+++ NEWBIE/TO-DO interactive-comments-section-main/JavaScript/card/commentary.js
@@ -1,15 +1,9 @@
-// import { currentUser } from "../comments.js";
-import { renderComments } from "../renderComments.js";
-import {
-  createElement,
-  createAvatar,
-  createBtn,
-  generateId,
-} from "../utilities-ui.js";
+import { renderComments, renderCommentsReply } from "../renderComments.js";
+import { createElement, createAvatar, createBtn } from "../utilities-ui.js";
 import createTextareaComment from "./contentCommentary.js";
 import action from "../data/acciones.js";
-import local from "../module/localStorage.js";
 import { commentLocal } from "../infoLocalStorage.js";
+import { createNewCommentary } from "./createNewComment.js";
 
 export const sectionAddComentary = (
   image,
@@ -17,22 +11,19 @@ export const sectionAddComentary = (
   actionBtn = action.reply,
   id
 ) => {
-  const imageCurrentUser = commentLocal.currentUser.image.png;
-  const nameUserReply = commentLocal.currentUser.username;
-
   const contentCommentary = createElement("div", "content-card");
+  sectionAddComentary.divCommentary = contentCommentary;
   const photoUser = createAvatar(image);
   const inputHidden = createElement("input");
   inputHidden.setAttribute("type", "hidden");
-  inputHidden.value = id;
-
+  // inputHidden.value = id;
+  sectionAddComentary.inputHidden = inputHidden;
   const aside = createElement("aside", "avatar-commentary");
   aside.append(photoUser);
   let txtBoton = action.reply;
 
   if (actionBtn == action.send) {
     txtBoton = action.send;
-    contentCommentary.classList.remove("hidden");
   }
   const buttonReply = createBtn(txtBoton);
 
@@ -59,35 +50,7 @@ export const sectionAddComentary = (
   });
 
   buttonReply.addEventListener("click", () => {
-    const newComment = {
-      id: generateId(),
-      content: txtComent.value,
-      createdAt: "JUST NOW",
-      score: 0,
-      replyingTo: username,
-      user: {
-        image: {
-          png: imageCurrentUser,
-          webp: imageCurrentUser,
-        },
-        username: nameUserReply,
-      },
-      replies: [],
-    };
-
-    const validationButtonReplyOSend = actionBtn == action.send ? false : true;
-
-    const cardReply = renderComentaryReply(
-      newComment,
-      validationButtonReplyOSend
-    );
-    contentCommentary.innerHTML = "";
-    contentCommentary.classList.remove("content-card");
-
-    contentCommentary.append(cardReply, inputHidden);
-
-    validation.disabled = true;
-    local.add("comentarios", newComment, inputHidden.value);
+    createNewCommentary(txtComent.value, username, actionBtn, validation, id);
   });
 
   return contentCommentary;
@@ -96,15 +59,10 @@ export const sectionAddComentary = (
 const renderComentary = (comentario) => {
   const { user } = comentario;
   let image, username;
-  // if (isReply) {
-  //   username = user.username;
-  // } else {
-  //   username = comentario;
-  // }
 
   if (user) {
     image = user.image.png;
-    username =  user.username;
+    username = user.username;
   }
   let contentCommentary = sectionAddComentary(
     commentLocal.currentUser.image.png,
@@ -113,10 +71,10 @@ const renderComentary = (comentario) => {
 
   return contentCommentary;
 };
-const renderComentaryReply = (comentario, isReply) => {
-  const cardReply = commentaryReply(comentario, isReply);
-  return cardReply;
-};
+// export const renderComentaryReply = (comentario, isReply) => {
+//   const cardReply = commentaryReply(comentario, isReply);
+//   return cardReply;
+// };
 
 export const commentaryReply = (comentario, isReply) => {
   const { id, content, user, createdAt } = comentario;
@@ -125,7 +83,7 @@ export const commentaryReply = (comentario, isReply) => {
   const replyCommentary = {
     id: id,
     content: content,
-
+    replica: true,
     createdAt,
     score: 0,
     replyingTo: username,
@@ -140,7 +98,7 @@ export const commentaryReply = (comentario, isReply) => {
   };
   commentaryReply.id = replyCommentary;
 
-  const renderComment = renderComments(replyCommentary, isReply);
+  const renderComment = renderCommentsReply(replyCommentary, isReply);
   if (isReply) renderComment.classList.add("reply-container");
 
   return renderComment;
