+++ NEWBIE/TO-DO interactive-comments-section-main/JavaScript/youProfile.js
@@ -1,0 +1,13 @@
+import { sectionAddComentary } from "./card/commentary.js";
+import action from "./data/acciones.js";
+import { commentLocal } from "./infoLocalStorage.js";
+export const mostrarMiPerfil = () => {
+  const contentHome = document.getElementById("content-home");
+  const youProfile = sectionAddComentary(
+    commentLocal.currentUser.image.png,
+    "commentLocal.currentUser.username",
+    action.send,
+    0
+  );
+  contentHome.append(youProfile);
+};
