+++ NEWBIE/TO-DO interactive-comments-section-main/styles/card.css
@@ -7,11 +7,15 @@
   width: 63rem;
 }
 
-.reply-container {
+.container-replies {
   margin-left: 3rem;
   border-left: solid 1px rgba(160, 160, 160, 0.699);
   padding-left: 5rem;
+  display: flex;
+  flex-direction: column;
+  gap: 3rem;
 }
+
 .content-card {
   display: flex;
   background-color: white;
